1.00.05

+++ commands/roll.js
@@ -5,7 +5,7 @@ module.exports = {
     permissions: [],
     description: "Roll a dice!",
     async execute(message, args, cmd, client, Discord, profileData) {
-        const randomNumber = Math.floor(Math.random() * args[0]) + 1;
+        var randomNumber = Math.floor(Math.random() * args[0]) + 1;
         const newEmbed = new Discord.MessageEmbed()
             .setColor('RANDOM')
             .setAuthor(`${message.author.username}`, `${message.author.displayAvatarURL()}`, '')
@@ -16,7 +16,7 @@ module.exports = {
             .setFooter(`Roll with ${prefix}roll or ${prefix}r`);
 
         if (!args[0]) return message.reply("Please enter a number");
-        if (isNaN(args[0])) return message.reply("Please enter a valid number");
+        if (isNaN(args[0]) || args[0] % 1 != 0 || args[0] <= 1) return message.reply("Please enter a valid whole number greater than 1");
         message.channel.send(newEmbed);
         message.delete();
     },
